requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.23.0
sounddevice>=0.4.6
soundfile>=0.12.1
evdev>=1.6.1